pyproject.toml
[project]
name = "televibecode"
version = "0.1.0"
description = "Remote orchestration harness for managing Claude Code sessions via Telegram"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "aiosqlite>=0.20",
    "mcp[cli]>=1.0",
    "python-telegram-bot>=21.0",
    "gitpython>=3.1",
    "structlog>=24.0",
    "pyyaml>=6.0",
    "aiofiles>=24.0",
    "python-dotenv>=1.0",
    "agno",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
televibecode = "televibecode.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/televibecode"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["UP017"]  # datetime.UTC not recognized by mypy, use timezone.utc

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = false
warn_unused_ignores = false
disallow_untyped_defs = false
ignore_missing_imports = true

# Per-module overrides for external library patterns
[[tool.mypy.overrides]]
module = "televibecode.telegram.*"
# telegram-bot handlers guarantee message/chat/callback exist in handlers
disable_error_code = ["union-attr", "no-any-return", "arg-type"]

[[tool.mypy.overrides]]
module = "televibecode.ai.*"
# agno doesn't have proper type stubs
disable_error_code = ["attr-defined", "import-untyped"]

[[tool.mypy.overrides]]
module = "televibecode.orchestrator.server"
# Database may be None before initialization but is checked at runtime
disable_error_code = ["arg-type", "union-attr"]

[[tool.mypy.overrides]]
module = "televibecode.config"
# pydantic-settings constructor patterns
disable_error_code = ["call-arg"]

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
]
